Add Code snippets for Contract component
