+++ src/api/hooks/useGetAccountResource.ts
@@ -3,6 +3,17 @@ import {useQuery, UseQueryResult} from "@tanstack/react-query";
 import {getAccountResource} from "..";
 import {ResponseError} from "../client";
 import {useGlobalState} from "../../global-config/GlobalConfig";
+import {orderBy} from "lodash";
+
+export type ModuleMetadata = {
+  name: string;
+  source: string;
+};
+
+export type PackageMetadata = {
+  name: string;
+  modules: ModuleMetadata[];
+};
 
 export function useGetAccountResource(
   address: string,
@@ -19,4 +30,23 @@ export function useGetAccountResource(
       ),
     {refetchOnWindowFocus: false},
   );
+}
+
+export function useGetAccountPackages(address: string) {
+  const {data: registry} = useGetAccountResource(
+    address,
+    "0x1::code::PackageRegistry",
+  );
+
+  const registryData = registry?.data as {
+    packages?: PackageMetadata[];
+  };
+
+  const packages: PackageMetadata[] =
+    registryData?.packages?.map((pkg): PackageMetadata => {
+      const sortedModules = orderBy(pkg.modules, "name");
+      return {name: pkg.name, modules: sortedModules};
+    }) || [];
+
+  return orderBy(packages, "name");
 }
