+++ src/pages/Account/Tabs/ModulesTab/Contract.tsx
@@ -1,5 +1,5 @@
 import {Types} from "aptos";
-import {ReactNode, useEffect, useState} from "react";
+import {ReactNode, useState} from "react";
 import Error from "../../Error";
 import {useGetAccountModules} from "../../../../api/hooks/useGetAccountModules";
 import EmptyTabContent from "../../../../components/IndividualPageContent/EmptyTabContent";
@@ -28,6 +28,11 @@ import {useGlobalState} from "../../../../global-config/GlobalConfig";
 import {view} from "../../../../api";
 import {grey} from "../../../../themes/colors/aptosColorPalette";
 import { useNavigate } from "../../../../routing";
+import {Code} from "../../Components/CodeSnippet";
+import {
+  PackageMetadata,
+  useGetAccountPackages,
+} from "../../../../api/hooks/useGetAccountResource";
 
 type ContractFormType = {
   typeArgs: string[];
@@ -46,6 +51,10 @@ function Contract({address, isRead}: {address: string; isRead: boolean}) {
   const isWideScreen = useMediaQuery(theme.breakpoints.up("md"));
   const {data, isLoading, error} = useGetAccountModules(address);
   const {selectedModuleName, selectedFnName} = useParams();
+  const sortedPackages: PackageMetadata[] = useGetAccountPackages(address);
+  const selectedModule = sortedPackages
+    .flatMap((pkg) => pkg.modules)
+    .find((module) => module.name === selectedModuleName);
 
   if (!isRead && !isWideScreen) {
     return (
@@ -141,19 +150,26 @@ function Contract({address, isRead}: {address: string; isRead: boolean}) {
         />
       </Grid>
       <Grid item md={9} xs={12}>
-        {!module || !fn ? (
-          <Box
-            padding={4}
-            bgcolor={theme.palette.mode === "dark" ? grey[800] : grey[100]}
-            borderRadius={1}
-          >
+        <Box
+          padding={4}
+          bgcolor={theme.palette.mode === "dark" ? grey[800] : grey[100]}
+          borderRadius={1}
+        >
+          {!module || !fn ? (
             <Typography>Please select a function</Typography>
-          </Box>
-        ) : isRead ? (
-          <ReadContractForm module={module} fn={fn} key={contractFormKey} />
-        ) : (
-          <RunContractForm module={module} fn={fn} key={contractFormKey} />
-        )}
+          ) : isRead ? (
+            <ReadContractForm module={module} fn={fn} key={contractFormKey} />
+          ) : (
+            <RunContractForm module={module} fn={fn} key={contractFormKey} />
+          )}
+
+          {module && fn && selectedModule && (
+            <>
+              <Divider sx={{margin: "16px 0"}} />
+              <Code bytecode={selectedModule?.source} />
+            </>
+          )}
+        </Box>
       </Grid>
     </Grid>
   );
@@ -175,9 +191,6 @@ function ContractSidebar({
       bgcolor={theme.palette.mode === "dark" ? grey[800] : grey[100]}
       borderRadius={1}
     >
-      <Typography fontSize={16} fontWeight={500} marginBottom={"24px"}>
-        Select function
-      </Typography>
       <Box>
         {Object.entries(moduleAndFnsGroup)
           .sort((a, b) => a[0].localeCompare(b[0]))
